Move Ember.TextArea to ember-views.
