+++ packages/ember-handlebars/lib/controls.js
@@ -1,6 +1,6 @@
 import Checkbox from "ember-views/views/checkbox";
 import TextField from "ember-views/views/text_field";
-import TextArea from "ember-handlebars/controls/text_area";
+import TextArea from "ember-views/views/text_area";
 
 import Ember from "ember-metal/core"; // Ember.assert
