+++ packages/ember-views/lib/views/text_area.js
